+++ examples/src/bin/dma_mem2mem.rs
@@ -7,7 +7,7 @@
 #![no_main]
 
 use esp_backtrace as _;
-use esp_hal::{delay::Delay, dma::Mem2Mem, dma_buffers, prelude::*};
+use esp_hal::{delay::Delay, dma::Mem2Mem, dma_buffers, entry, time::ExtU64};
 use log::{error, info};
 
 const DATA_SIZE: usize = 1024 * 10;